requirements.txt
pytest==8.4.1
selene>=2.0.b4
allure-pytest
python-dotenv
allure-python-commons~=2.15.0
selenium>=4.12.0
PyPDF2